ship acceleration

+++ COMP2000-Project/src/Planet.java
@@ -1,7 +1,7 @@
 public class Planet extends SimulationObject {
     public int gravityStrength;
 
-    public Planet (int mass, int size, int x, int y, int gravityStrength){
+    public Planet (float mass, int size, int x, int y, int gravityStrength){
         super(mass, size, x, y);
         this.gravityStrength = gravityStrength;
     }
